# Improve context cancellation performance.