See: https://github.com/ncruces/go-sqlite3/discussions/247#discussioncomment-12549316

This currently just adds a (micro) benchmark to measure the problem.

Results:
```
goos: linux
goarch: amd64
pkg: github.com/ncruces/go-sqlite3/driver
cpu: Intel(R) Xeon(R) W-2135 CPU @ 3.70GHz
         │  1_off.txt  │          2_noctx.txt          │           3_nosched.txt            │               4_tip.txt                │              5_1000.txt              │
         │   sec/op    │   sec/op     vs base          │   sec/op     vs base               │    sec/op      vs base                 │    sec/op     vs base                │
_loop-12   807.8m ± 2%   817.3m ± 2%  ~ (p=0.190 n=10)   828.9m ± 2%  +2.61% (p=0.015 n=10)   2084.9m ± 13%  +158.09% (p=0.000 n=10)   897.4m ± 10%  +11.08% (p=0.000 n=10)
```

Interpreting results:
- `off` is disabling progress reporting by commenting this [`c.call("sqlite3_progress_handler_go", ...)`](https://github.com/ncruces/go-sqlite3/blob/v0.24.1/conn.go#L123)
- `noctx` is passing an non-cancellable context, which skips much of [`SetInterrupt`](https://github.com/ncruces/go-sqlite3/blob/v0.24.1/conn.go#L342), and [`Gosched`, with cheap `Done`/`Err` implementations](https://github.com/ncruces/go-sqlite3/blob/v0.24.1/conn.go#L377-L387)
- `nosched` just comments `Gosched` (but there are still mutexes/atomics on `Done`/`Err`)
- `tip` is where we currently stand with a cancellable context
- `1000` is changing the frequency of the callback from 100 to 1000