@ncruces 2025-03-21
New implementation results:
```
goos: linux
goarch: amd64
pkg: github.com/ncruces/go-sqlite3/driver
cpu: Intel(R) Xeon(R) W-2135 CPU @ 3.70GHz
         │    tip.txt    │               off.txt               │               new.txt               │
         │    sec/op     │   sec/op     vs base                │   sec/op     vs base                │
_loop-12   2084.9m ± 13%   807.8m ± 2%  -61.25% (p=0.000 n=10)   805.0m ± 2%  -61.39% (p=0.000 n=10)
```