@ncruces 2025-03-21
https://github.com/golang/go/issues/64417 makes it sound that calling back every few hundreds of μs is OK; a value of 1000 should be OK. And since the big cost is `Gosched` we avoid doing _that_ every time.

Combined, this means that we're calling `Gosched` every 16000 VDBE opcodes, or roughly every 10ms (compiler), probably around every 100ms (interpreter), regardless of there being a context. `wasip1` seems OK with this, no regressions.